create hook outside of component
