+++ src/components/ListOfPhotoCards/index.js
@@ -4,7 +4,7 @@ import { PhotoCard } from '../PhotoCard';
 export const ListOfPhotoCards = () => {
   return (
   <ul>
-    {[1, 2, 3, 5, 6, 7].map(id => <PhotoCard key={id} />)}
+    {[1, 2, 3, 5, 6, 7].map(id => <PhotoCard key={id} id={id} />)}
   </ul>
   )
 }
