+++ src/components/PhotoCard/index.js
@@ -1,50 +1,19 @@
-import React, { Fragment, useEffect, useRef, useState } from 'react';
+import React, { Fragment,} from 'react';
 import { Article, ImgWrapper, Img, Button } from './styles';
 import { MdFavoriteBorder, MdFavorite } from 'react-icons/md';
+import { useLocalStorage } from '../../hooks/useLocalStorage';
+import { useNearScreen } from '../../hooks/useNearScreen';
 
 const DEFAULT_IMAGE = 'https://images.unsplash.com/photo-1518791841217-8f162f1e1131?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60'
 
 export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
-  const element = useRef(null)
-  const [show, setShow] = useState(false)
+  const [show, element] = useNearScreen()
   const key = `like-${id}`
+  const [liked, setLiked] = useLocalStorage(key, false)
 
-  const [liked, setLiked] = useState(() =>{
-    try {
-      const like = window.localStorage.getItem(key)
-      return like
-    } catch (e) {
-    return false
-    }
-  })
-
-  console.log(liked)
-  useEffect(function () {
-    Promise.resolve(
-      typeof window.IntersectionObserver !== 'undefined'
-        ? window.IntersectionObserver
-        : import('intersection-observer')
-    ).then(() => {
-      const observer = new window.IntersectionObserver(function (entries) {
-        const { isIntersecting } = entries[0]
-        if (isIntersecting) {
-          setShow(true)
-          observer.disconnect()
-        }
-      })
-      observer.observe(element.current)
-    })
-  }, [element])
 
   const Icon = liked ? MdFavorite : MdFavoriteBorder
-  const setLocalStorage = value =>{
-    try {
-      window.localStorage.setItem(key, value)
-      setLiked(value)
-    } catch (e) {
-      console.log(e)
-    }
-  }
+
   return (
     <Article ref={element}>
       {
@@ -55,7 +24,7 @@ export const PhotoCard = ({ id, likes = 0, src = DEFAULT_IMAGE }) => {
             </ImgWrapper>
           </a>
 
-          <Button onClick={()=> setLocalStorage(!liked)}>
+          <Button onClick={()=> setLiked(!liked)}>
             <Icon size='32px' /> {likes} likes!
           </Button>
         </Fragment>
