+++ src/hooks/useLocalStorage.js
@@ -1,0 +1,24 @@
+import { useState } from 'react';
+
+
+export function useLocalStorage (key, initialValue) {
+  const [storedValue, setValue] = useState(() =>{
+    try {
+      const item = window.localStorage.getItem(key)
+      return item !== null ? JSON.parse(item) : initialValue
+    } catch (e) {
+    return initialValue
+    }
+  })
+
+  const setLocalStorage = value =>{
+    try {
+      window.localStorage.setItem(key, JSON.stringify(value))
+      setValue(value)
+    } catch (e) {
+      console.log(e)
+    }
+  }
+  return [storedValue, setLocalStorage]
+
+}
